normalisation/requirements.txt
fastapi
uvicorn
pandas
numpy
google-cloud-storage
scikit-learn
pyarrow
openpyxl
scipy
